Refactor for polymorphic parsing
