+++ log-viewer/jest.config.js
@@ -1,7 +1,7 @@
 /** @type {import('ts-jest/dist/types').InitialOptionsTsJest} */
 module.exports = {
   preset: "ts-jest",
-  testEnvironment: "node",
+  testEnvironment: "jsdom",
   transform: {
     // transform files with ts-jest
     "^.+\\.(js|ts)$": "ts-jest",
