+++ log-viewer/package.json
@@ -22,18 +22,19 @@
 		"@rollup/plugin-commonjs": "^20.0.0",
 		"@rollup/plugin-node-resolve": "^13.0.4",
 		"@rollup/plugin-typescript": "^8.2.1",
-		"@types/jest": "^27.0.2",
+		"@types/jest": "^27.5.2",
 		"@typescript-eslint/eslint-plugin": "^4.28.5",
 		"@typescript-eslint/parser": "^4.28.5",
 		"babel-plugin-rewire": "^1.2.0",
 		"concurrently": "^6.2.1",
 		"http-server": "^13.0.2",
-		"jest": "^27.3.1",
+		"jest": "^28.1.2",
+		"jest-environment-jsdom": "^28.1.2",
 		"postcss": "^8.3.5",
 		"rollup": "^2.0.0",
 		"rollup-plugin-postcss": "^4.0.0",
 		"rollup-plugin-terser": "^7.0.2",
-		"ts-jest": "^27.0.7",
+		"ts-jest": "^28.0.5",
 		"tslib": "^2.3.0",
 		"typescript": "^4.0.0"
 	}
